requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.9.2
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
Django==5.2.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.127.1
greenlet==3.3.0
h11==0.16.0
idna==3.11
image==1.5.33
passlib==1.7.4
pillow==11.3.0
py==1.11.0
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-jose==3.5.0
python-multipart==0.0.21
qrcode==8.2
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
sqlparse==0.5.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.40.0
